+++ packages/orq-rc/RELEASES.md
@@ -2499,3 +2499,13 @@ Based on:
 - [typescript v3.10.0-rc.6] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.6 - packages/orq-rc
+
+## 2025-07-14 14:57:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.581.1 (2.656.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.7] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.7 - packages/orq-rc